Completed Second Section and Outline
This update added an image file, built a complete outline of the webpage
in html, and completed the second section of the webpage. Project is
halfway done, next update is to finish up the third section and complete
the last stretch if time permits.
-Munchnax

Added file(s):
- pokeball-16x16.png (created by me)

This update completed:
- the outline of the webpage (text/images) and added CSS selectors
- Spacing and styling of text and images for section 2