Completed the Webpage
Finished section 3, section 4, and the footer.

Note: This will probably be the last commit to this project. Thanks to
The Odin Project for the project assignment! This assignment helped me
put my vague and confusing understanding of CSS, Flexbox, and HTMl and
prove to myself I understand and can use it properly.

-Munchnax